login, logout (jwt), creat model class

+++ TimeSheetProject/src/main/java/com/timesheet/timesheetproject/dto/response/ApiResponse.java
@@ -11,7 +11,7 @@ import lombok.experimental.FieldDefaults;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ApiResponse <T>{
-    int code = 1000;
+    long code = 1000;
     String message;
     T result;
 }
